\docType{package}
\name{pisa-package}
\alias{pisa}
\alias{pisa-package}
\title{Programme for International Student Assessment (PISA)}
\description{
  Programme for International Student Assessment (PISA)
}
\details{
  This package provides the 2009 PISA data formatted for R.
}
\author{
  Jason Bryer \email{jason@bryer.org}
}
\references{
  Organisation for Economic Co-operation and Development
  (2009).  Programme for International Student Assessment
  (PISA).  \url{http://www.pisa.oecd.org/}
}
\keyword{data}
\keyword{datasets}

