\docType{data}
\name{pisa.catalog.students}
\alias{pisa.catalog.students}
\title{Data catalog for student file.}
\format{a named character vector where name is the variable name and the value
        the description.}
\source{
  Organization for Economic Co-operation and Development
}
\description{
  Data catalog for student file.
}
\references{
  Organisation for Economic Co-operation and Development
  (2009).  Programme for International Student Assessment
  (PISA).  \url{http://www.pisa.oecd.org/}
}
\keyword{datasets}

