\docType{data}
\name{pisa.countries}
\alias{pisa.countries}
\title{Data frame containing two columns, one with three letter country abbreviation
and the other with the full country name.}
\format{a data frame with 65 ovservations of 2 variables.}
\source{
  Organization for Economic Co-operation and Development
}
\description{
  Data frame containing two columns, one with three letter
  country abbreviation and the other with the full country
  name.
}
\references{
  Organisation for Economic Co-operation and Development
  (2009).  Programme for International Student Assessment
  (PISA).  \url{http://www.pisa.oecd.org/}
}
\keyword{datasets}

