\docType{data}
\name{pisa.parent}
\alias{pisa.parent}
\title{Parent survey results of the 2009 Programm of International Student Assessment.}
\format{a data frame}
\source{
  Organization for Economic Co-operation and Development
}
\description{
  Parent survey results from the 2009 Programme of
  International Student Assessment (PISA) as provided by
  the Organization for Economic Co-operation and
  Development (OECD). See \url{http://www.pisa.oecd.org/}
  for more information including the code book.
}
\references{
  Organisation for Economic Co-operation and Development
  (2009).  Programme for International Student Assessment
  (PISA).  \url{http://www.pisa.oecd.org/}
}
\keyword{datasets}

