\name{recodeSudentPISA}
\alias{recodeSudentPISA}
\title{Recodes the factor variables variables within PISA.}
\usage{
  recodeSudentPISA(x)
}
\arguments{
  \item{x}{data frame from PISA}
}
\value{
  data frame with variables recoded as factors.
}
\description{
  Recodes the factor variables variables within PISA.
}

