\name{spss.parse.labels}
\alias{spss.parse.labels}
\title{Modified from the memisc package.
http://cran.r-project.org/web/packages/memisc}
\usage{
  spss.parse.labels(file)
}
\arguments{
  \item{file}{the SPSS syntax file}
}
\value{
  value labels
}
\description{
  Modified from the memisc package.
  http://cran.r-project.org/web/packages/memisc
}

