\name{spss.parse.variable.labels}
\alias{spss.parse.variable.labels}
\title{Copied from the memisc package
http://cran.r-project.org/web/packages/memisc}
\usage{
  spss.parse.variable.labels(file)
}
\arguments{
  \item{file}{the SPSS syntax file}
}
\value{
  variable labels
}
\description{
  Copied from the memisc package
  http://cran.r-project.org/web/packages/memisc
}

